move: AuthButtonをauthディレクトリにしまう
