+++ src/components/auth/LoginForm.tsx
@@ -1,5 +1,5 @@
 import { Box, Typography, Divider, Stack } from "@mui/material";
-import { AuthButton } from "../shared/button";
+import { AuthButton } from "./AuthButton";
 import { FaDiscord } from "react-icons/fa";
 import { FcGoogle } from "react-icons/fc";
 import { signIn } from "next-auth/react";
