+++ src/components/shared/button/index.ts
@@ -1,3 +1,2 @@
-export { AuthButton } from "./AuthButton";
 export { Button } from "./Button";
 export { PostNavigationButton } from "./PostNavigationButton";
